Support additional CMake build options
- Add option BUILD_TESTS. Default to ON (following existing behavior)
- Add option BUILD_README_HELPER. Default to ON (following existing behavior).
- Add option BUILD_SHARED_LIBS. Default to OFF (following existing behavior).
- Add option USE_EXTERNAL_JSON. Default to OFF (following existing behavior).

BUILD_TESTS and BUILD_README_HELPER toggle whether contents in subdirectories
"test" and "readme-helper" will be built.

BUILD_SHARED_LIBS toggles whether libsdptransform will be built as a static
or shared library. add_library automatically checks the value of
BUILD_SHARED_LIBS at configure/build time.

USE_EXTERNAL_JSON toggles whether an external (e.g. system) copy of
nlohmann-json should be used instead of the bundled one. If enabled,
CMake will automatically search for an installation of nlohmann-json on the
system and configure step will fail if not found. If nlohmann-json is
installed in a non-standard location, CMAKE_MODULE_PATH can be set so that
CMake can find the nlohmann-json installation.

json.hpp has been moved from include/ to new folder thirdparty/, so that
the thirdparty folder can be added as an include directory when external
nlohmann-json is not in use.

include/sdptransform.hpp is now include/sdptransform.hpp.in. At configure
time, the .hpp.in file is preprocessed into the build tree. This step is
necessary in order to set the #include path to json.hpp in sdptransform.hpp.